% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIso3166.R
\name{convert_iso3166}
\alias{convert_iso3166}
\title{Convert answer codes}
\usage{
convert_iso3166(from.acode, from.alang, iso3166map)
}
\arguments{
\item{from.acode}{Specifies the answer code to be converted. Usually a vector of variable (column) which 
stores answer codes regarding nationality or country.}

\item{from.alang}{Specifies the specific language of the labels to be converted. This is synonymous to the 
language of the submission (startlanguage)}

\item{iso3166map}{Data frame which maps answer codes from one language to the other. @seealso \code{\link{map_iso3166}}.}
}
\value{
Vector of answer codes in the target language.
}
\description{
Convert actual answer codes in a result data set from one language to the other. Requires a 
 mapping table produced by \code{\link{map_iso3166}}. If an answer code does not match, the original code will 
 be returned.
}
