% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_docx.R
\name{export_docx}
\alias{export_docx}
\title{Export Labels to Word}
\usage{
export_docx(file, plang)
}
\arguments{
\item{file}{String specifying path and file name}

\item{plang}{String language code.}
}
\value{
Nothing. Writes file to disc.
}
\description{
Question and answer labels can be exported as word document (docx). Only contains basic formatting.
}
