% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractQuestionLabels.R
\name{extract_question_labels}
\alias{extract_question_labels}
\title{Extract Question Labels}
\usage{
extract_question_labels(file, strip_html = T)
}
\arguments{
\item{file}{char path to exported LimeSurvey Survey Structure *.lss file}

\item{strip_html}{logical. Default set to \code{TRUE} will remove any simple HTML tags from the labels. Otherwise leaves 
label string as exported from LS.}
}
\value{
data frame containing the following fields:
 \itemize{
   \item{qid - question with id}
   \item{qcode - LS question code, e.g. SDEM001, GlickM01, }
   \item{subqcode - LS sub-question code, e.g. SQ001, SQ002, etc. }
   \item{qtxt - LS question text}
   \item{subqtxt - LS sub-question text}
   \item{onecode - question and subquestion code as one string}
   \item{lang - language of answer text}
 }
}
\description{
Extract all question labels from a LimeSurvey Survey Structure file (*.lss). The *.lss file is
 exported manually and downloaded. This function extracts all question and sub-question texts in
 all available languages and returns them as data frame.
}
