% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResponseLabels.R
\name{extract_response_labels}
\alias{extract_response_labels}
\title{Extract Response Labels}
\usage{
extract_response_labels(file)
}
\arguments{
\item{file}{char path to exported LimeSurvey Survey Structure *.lss file}
}
\value{
data frame containing the following fields:
 \itemize{
   \item{qid - answer correspond to question with id}
   \item{aid - answer id}
   \item{acode - LS answer code}
   \item{atxt - LS answer text}
   \item{lang - language of answer text}
 }
}
\description{
Extract all response labels from a LimeSurvey Survey Structure file (*.lss). The *.lss file is
 exported manually and downloaded. This function extracts all answer codes, their corresponding answer texts in
 all available languages and returns them as data frame.
}
