% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getI18n.R
\name{get_i18nx}
\alias{get_i18nx}
\title{Get translation of label}
\usage{
get_i18nx(othcode, plangcode)
}
\arguments{
\item{othcode}{String. The message key.}

\item{plang}{String. Language code}
}
\value{
Translated label if code and language exists. English if language not available, or code if 
 language does not exist.
}
\description{
Hack to retrieve the translated "other" label. TODO: should be done via base::gettext
}
