% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabel.R
\name{get_lsLabel}
\alias{get_lsLabel}
\title{Get Question Label}
\usage{
get_lsLabel(x)
}
\arguments{
\item{x}{Variable column of the result data frame}
}
\value{
String of question label or \code{NULL} if no label has been set.
}
\description{
Retrieves question or sub-question text used in a LimeSurvey questionnaire. Implies that the question label 
 has been previously assigned by calling the \code{\link{set_question_labels}} function. The question label is stored 
 as attribute "label" for each column of the data frame. 
 
 In case of sub-questions, the label of the sub-question item is returned, not the text of the overarching (parent) 
 question. To retrieve the corresponding parent question see \code{\link{get_lsParentlabel}}.
}
