% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParentLabel.R
\name{get_lsParentlabel}
\alias{get_lsParentlabel}
\title{Get Parent Question Label}
\usage{
get_lsParentlabel(x)
}
\arguments{
\item{x}{Variable column of the result data frame}
}
\value{
String of parent question label or \code{NULL} if no parent label has been set.
}
\description{
Retrieves parent question label for sub-questions of a LimeSurvey questionnaire. Implies that the question label 
 has been previously assigned by calling the \code{\link{set_question_labels}} function. The parent question label is stored 
 as attribute "parent_label" for each sub-question variable of the data frame.
}
