% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSubquestion.R
\name{is_lsSubquestion}
\alias{is_lsSubquestion}
\title{Checks if variable is of sub-question}
\usage{
is_lsSubquestion(x)
}
\arguments{
\item{x}{Variable column of the result data frame}
}
\value{
Bool. \code{TRUE} in case the column holds responses of a sub-question
}
\description{
Columns in LimeSurvey based result data frames hold variables that can represent questions or
 sub-questions. In case a variable represents a sub-question, it can be useful to call \code{\link{get_lsParentlabel}} 
 in order to retrieve the text of the overarching question. 
 
 The function simply checks if the parent label is NULL or not.
}
