% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{read_gpg}
\alias{read_gpg}
\title{Read GPG encrypted data}
\usage{
read_gpg(file, as_text = FALSE)
}
\arguments{
\item{file}{String. File path of file to be decrypted}
}
\value{
R data frame
}
\description{
Read and deserialize GPG encrypted file. Requires
 private key for decrypting file. Will prompt for password of private key.
}
