% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setQuestionLabels.R
\name{set_question_labels}
\alias{set_question_labels}
\title{Set variable label on exported LS result data}
\usage{
set_question_labels(data, labels, plang)
}
\arguments{
\item{data}{Data frame of LimeSurvey R result data set.}

\item{labels}{Data frame extracted from LimeSurvey *.lss xml file. See \code{\link{extract_question_labels}}}

\item{plang}{char language code. One of the available language codes from the survey}
}
\value{
data frame with attached text label for each variable (question or sub-question).
}
\description{
LimeSurvey offers the option to export survey result data as an R data frame. Column names are question
 codes while variable values are response codes. This function assigns the question texts in the selected
 language to each variable (column). The question text is available as attribute "label" for each column. Useful in
 conjunction with \code{sjlabelled::get_label()} or \code{ggplot} which can automatically retrieves variable
 labels for graphics and/or frequency tables.
}
\seealso{
\code{\link{extract_question_labels}}
}
