% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{write_gpg}
\alias{write_gpg}
\title{Write GPG encrypted data}
\usage{
write_gpg(df, file = "df.gpg", receiver)
}
\arguments{
\item{df}{The data frame to be encrypted}

\item{file}{String. Name and path for writing file}

\item{receiver}{email address of receiver (which has gpg key associated)}
}
\value{
Nothing
}
\description{
Utility function to serialize and store data frame as encrypted file on disc.  Requires
 a valid gpg public key.
}
