% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setResponseLabels.R
\name{set_response_labels}
\alias{set_response_labels}
\title{Set response labels on exported LS result data}
\usage{
set_response_labels(data, labels, plang)
}
\arguments{
\item{data}{Data frame of LimeSurvey R result data}

\item{labels}{Data frame extracted from LimeSurvey *.lss xml file. See \code{\link{extract_response_labels}}}

\item{lang}{char language code}
}
\value{
Data frame with value (answer) labels assigned to each variable (column).
}
\description{
LimeSurvey offers the option to export survey result data as an R data frame. Variable values are
 stored as response codes. This function assigns value labels for each variable. Overall, for most variable types
 this means to convert the corresponding column and its values to a \code{factor} variable. Except for numberic,
 date and different open text fields (short, middle and long), all LimeSurvey and hence question types are converted
 to \code{factor}. LimeSurvey response codes are converted to factor levels while LimeSurvey response texts are assigned
 to the corresponding factor labels.
}
\details{
Answer labels that are "Other" with text field have their own variable names of the form "questioncode.other" and
question type "!" (dropdown)
}
\seealso{
\code{\link{extract_response_labels}}
}
